repository cleanup and refactor: changed folder structure for app
