+++ blockchain/interface/join-pool-data.interface.ts
