+++ blockchain/interface/login-context.interface.ts
@@ -1,0 +1,10 @@
+import { IPool } from "@/blockchain/interface/pool.interface";
+
+export interface ILogInContextType {
+  account: string | null;
+  signer: any; // replace 'any' with the actual type
+  connectProvider: () => Promise<void>;
+  disconnectProvider: () => Promise<void>;
+  existingPools: Array<IPool>;
+  setExistingPools: (pools: Array<IPool>) => void;
+}
