+++ blockchain/model/pool-join-data.ts
